% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvDFA.R
\name{mvDFA}
\alias{mvDFA}
\title{Analyse multivariate correlated time series and estimate long memory}
\usage{
mvDFA(
  X,
  steps = 50,
  degree = 1,
  verbose = F,
  cores = 1,
  covlist = F,
  brownian = F
)
}
\arguments{
\item{X}{Matrix or data.frame containing the time series in long format.}

\item{steps}{Maximum number of window sizes. These are spread logarithmically. If time series is short and steps is large, fewer window sizes are drawn. Default to `50`.}

\item{degree}{The maximum order of the detrending polynomial in the segments. This influences the smallest window size "minS" such that minS = degree + 2.}

\item{verbose}{Indicator whether additional infos should be printed. Default to `TRUE`.}

\item{cores}{Number of cores used in computation. Default to `1`.}

\item{covlist}{Indicator whether covariance of the time series per window size should be saved in a list.}

\item{brownian}{Indicator whether time series are assumed to be brownian (i.e. variance increases proportional to time)}
}
\value{
Returns list of Root Mean Squares per window size for the total `RMS_tot`, the generalized `RMS_gen`, the covariance approach `Cov_RMS_s` and the eigen values `RMS_eigen`,  the window sizes `S`, the estimated long memory coefficient for the multivariate time series using the total variance approach `Ltot`, the generalied approach `Lgen`, the average covariance approach `Lfull` and the eigen values approach `Leigen`. Further a list of covariance matrices per `S` may be returned. `MLE_H_element_wise` is the elementwise Hurst exponent estimated via `mleHK` of the `HKprocess` package for comparison.
}
\description{
Analyse multivariate correlated time series and estimate long memory
}
