% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvDFA.R
\name{print.mvDFA}
\alias{print.mvDFA}
\title{print object of class mvDFA}
\usage{
\method{print}{mvDFA}(x, ...)
}
\arguments{
\item{x}{object of class DFA to print.}

\item{...}{further parameters.}
}
\value{
Truncates the output printed into the console of objects of class \code{mvDFA}, but does not change object itself.
}
\description{
print object of class mvDFA
}
