% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lavMTS.R
\name{simulate_lavMTS}
\alias{simulate_lavMTS}
\title{Approximate correlated time series with common Hurst Exponent using a latent variable approach}
\usage{
simulate_lavMTS(N, H, Lambda, nxi = 1, Phi = NULL, d = 3, H_residuals)
}
\arguments{
\item{N}{Length of Times Series}

\item{H}{Vector of Hurst Exponents for latent variables. These are then mixed using the Cholesky decomposition of the given correlation matrix Phi.}

\item{Lambda}{Factor loadings of latent variable approach of dimension `d` times `nxi`.}

\item{nxi}{Number of latent variables used to generate time series.}

\item{Phi}{Correlation matrix of latent variables used to generate different Hurst Exponents.}

\item{d}{Dimension of time series.}

\item{H_residuals}{Vector of Hurst Exponents of the variable specific residuals (not due to the latent variable).}
}
\value{
Returns list of class "MultiTimeseriesLongMemory" containing `X` the multivariate time series as well as the model implied correlation matrix `implied_cor`.  The Hurst exponents are only approximate univariatly, since they result from mixed time series. Uncorrelated time series keep their univariate Hurst exponent `H`.
}
\description{
Approximate correlated time series with common Hurst Exponent using a latent variable approach
}
