% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Lorenz_noise.R
\name{simulate_Lorenz_noise}
\alias{simulate_Lorenz_noise}
\title{Simulate the Lorenz System with noise}
\usage{
simulate_Lorenz_noise(
  N = 1000,
  delta_t = NULL,
  tmax = 50,
  X0 = 0,
  Y0 = 1,
  Z0 = 1,
  sdX = NULL,
  sdY = NULL,
  sdZ = NULL,
  sdnoiseX,
  sdnoiseY,
  sdnoiseZ,
  s = 10,
  r = 28,
  b = 8/3,
  return_time = T
)
}
\arguments{
\item{N}{Length of Times Series}

\item{delta_t}{Step size for time scale. If \code{NULL} this is derived using \code{N} and \code{tmax}. \code{DEFAULT} to \code{NULL}.}

\item{tmax}{Upper bound of the time scale. This argument is ignored if \code{delta_t} is provided. \code{DEFAULT} to \code{50}.}

\item{X0}{Initial value for X at t=0. \code{DEFAULT} to 0.}

\item{Y0}{Initial value for Y at t=0. \code{DEFAULT} to 1.}

\item{Z0}{Initial value for Z at t=0. \code{DEFAULT} to 1.}

\item{sdX}{Use this argument to rescale the X-coordinate to have the desired standard deviation (exactly). This is ignored if set to \code{NULL}. \code{DEFAULT} to \code{NULL}.}

\item{sdY}{Use this argument to rescale the Y-coordinate to have the desired standard deviation (exactly). This is ignored if set to \code{NULL}. \code{DEFAULT} to \code{NULL}.}

\item{sdZ}{Use this argument to rescale the Z-coordinate to have the desired standard deviation (exactly). This is ignored if set to \code{NULL}. \code{DEFAULT} to \code{NULL}.}

\item{sdnoiseX}{Standard deviation of Gaussian noise of X-coordinate.}

\item{sdnoiseY}{Standard deviation of Gaussian noise of Y-coordinate.}

\item{sdnoiseZ}{Standard deviation of Gaussian noise of Z-coordinate.}

\item{s}{s-parameter of the Lorenz ODE. See Vigniette for further details. \code{DEFAULT} to \code{10}, which is the original value chosen by Lorenz.}

\item{r}{r-parameter of the Lorenz ODE. See Vigniette for further details. \code{DEFAULT} to \code{28}, which is the original value chosen by Lorenz.}

\item{b}{b-parameter of the Lorenz ODE. See Vigniette for further details. \code{DEFAULT} to \code{8/3}, which is the original value chosen by Lorenz.}

\item{return_time}{Logical whether the time-coordinate should be included in the returned \code{data.frame}. \code{DEFAULT} to \code{TRUE}.}
}
\description{
Simulate the Lorenz System with noise
}
