% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvDFA.R
\name{mvDFA}
\alias{mvDFA}
\title{Analyse multivariate correlated time series and estimate long memory}
\usage{
mvDFA(
  X,
  steps = 50,
  degree = 1,
  verbose = F,
  cores = 1,
  covlist = F,
  brownian = F
)
}
\arguments{
\item{X}{Matrix or data.frame containing the time series in long format.}

\item{steps}{Maximum number of window sizes. These are spread logarithmically. If time series is short and steps is large, fewer window sizes are drawn. Default to `50`. The dimensions (`ncol(X)`) and the `degree` influence the smallest possible window size.}

\item{degree}{The maximum order of the detrending polynomial in the segments. This influences the smallest window size "minS" such that minS = degree + 2.}

\item{verbose}{Indicator whether additional infos should be printed. Default to `TRUE`.}

\item{cores}{Number of cores used in computation. Default to `1`.}

\item{covlist}{Indicator whether covariance of the time series per window size should be saved in a list.}

\item{brownian}{Indicator whether time series are assumed to be brownian (i.e. variance increases proportional to time)}
}
\value{
Returns list of Root Mean Squares per window size for the total `RMS_tot`, the generalized `RMS_gen`, and the covariance approach `Cov_RMS_s`,  the window sizes `S`, the estimated long memory coefficient for the multivariate time series using the total variance approach `Ltot`, and the generalized approach `Lgen`, the average covariance approach `Lfull`. Further a list of covariance matrices per `S` may be returned.
}
\description{
Analyse multivariate correlated time series and estimate long memory
}
\examples{
Sigma <- matrix(.5, 3, 3); diag(Sigma) <- 1
X <- mvtnorm::rmvnorm(n = 10^3, sigma = Sigma) # generate correlated white noise (i.i.d. multivariate normal variables)
mvDFA(X = X)
}
