% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DFA.R
\name{print.DFA}
\alias{print.DFA}
\title{print object of class DFA}
\usage{
\method{print}{DFA}(x, ...)
}
\arguments{
\item{x}{object of class DFA to print.}

\item{...}{further parameters.}
}
\value{
Truncates the output printed into the console of objects of class \code{DFA}, but does not change object itself.
}
\description{
print object of class DFA
}
