% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvDFA.R
\name{print.mvDFA}
\alias{print.mvDFA}
\title{print object of class DFA}
\usage{
\method{print}{mvDFA}(x, ...)
}
\arguments{
\item{x}{object of class DFA to print.}

\item{...}{further parameters.}
}
\description{
print object of class DFA
}
