% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cMTS.R
\name{simulate_cMTS}
\alias{simulate_cMTS}
\title{Approximate correlated time series with given Hurst Exponent}
\usage{
simulate_cMTS(N, H, Sigma)
}
\arguments{
\item{N}{Length of Times Series}

\item{H}{Hurst Exponents for `d` time series. These are then mixed using the Cholesky decomposition of the given covariance matrix `Sigma`.}

\item{Sigma}{Positive semi definite covariance matrix of desired multi dimensional time series.}
}
\value{
Returns a multivariate correlated time series with covariance matrix `Sigma`. The Hurst exponents are only approximate univariatly, since they result from mixed time series. Uncorrelated time series keep their univariate Hurst exponent `H`.
}
\description{
Approximate correlated time series with given Hurst Exponent
}
