% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_MTS.R
\name{simulate_MTS_mixed_white_pink_brown}
\alias{simulate_MTS_mixed_white_pink_brown}
\title{Approximate correlated time series from white, pink and brown noise}
\usage{
simulate_MTS_mixed_white_pink_brown(N, Sigma, process = "white")
}
\arguments{
\item{N}{Length of multivariate Times Series}

\item{Sigma}{Positive semi definite covariance matrix the increments of desired multi dimensional time series. The dimensionality of Sigma sets the dimension of the time series. The variance scale the time. If the variances are all 1, then each data point represents one unit of time.}

\item{process}{Type of process. Can either be "white", "brown" or "pink". Default to "white". If process is a vector, a mixture of the three process is generated, correlated by Sigma.}
}
\value{
Returns a multivariate correlated time series with covariance matrix `Sigma`. The Hurst exponents are only approximate univariatly, since they result from mixed time series. Here, a mixture of "white", "pink" and "brown" noise can be chosen from. Uncorrelated time series keep their univariate Hurst exponent `H`.
}
\description{
Approximate correlated time series from white, pink and brown noise
}
